Computer Vision Algorithms using MATLAB: EdgeDetection: We have a set of images and perform some edge detection techniques using some masks such as Average mask, gaussian mask, DoG masks, Laplace masks etc.
hybridknn : This is performing a kNN algorithm for segmentation and
grouping after some initial thresholding. We are looking for RGB
colours and based on the intensities we pass, it filters them using
thresholding. Then it performs kmeans to group the filtered image. Aim
is that the region of interest is only passed through and hence kmeans
will be more efficient,

kmeans : This algorithm used  k means learning technique to  group the
most similar pixels (based on intensities) together. Colour with the
same intensities are grouped together and assigned the same color.

MidLevelBiological: This is used to imitate the way the human brain
performs mid level vision.

region growing: Segmentation using region growing

thresholding:  basic thresholding algorithm. Filters the image based on
the intensities.